+++ lib/stoyco_shared.dart
@@ -2,7 +2,10 @@ library stoyco_shared;
 
 // Cache system
 export 'package:stoyco_shared/cache/cache_manager.dart';
+export 'package:stoyco_shared/cache/cache_utils.dart';
+export 'package:stoyco_shared/cache/global_cache_manager.dart';
 export 'package:stoyco_shared/cache/in_memory_cache_manager.dart';
+export 'package:stoyco_shared/cache/persistent_cache_manager.dart';
 export 'package:stoyco_shared/cache/repository_cache_mixin.dart';
 export 'package:stoyco_shared/cache/models/cache_entry.dart';
 
@@ -18,7 +21,6 @@ export 'package:stoyco_shared/notification/notification.dart';
 export 'package:stoyco_shared/partner/partner.dart';
 export 'package:stoyco_shared/news/models/feed_content_model.dart';
 export 'package:stoyco_shared/widgets/interactive_content_card.dart';
-export 'package:stoyco_shared/widgets/reactive_social_buttons.dart';
 export 'package:stoyco_shared/widgets/social_button.dart';
 export 'package:stoyco_shared/widgets/shared_like.dart';
 export 'package:stoyco_shared/widgets/loop_video_player.dart';